feat: move recommendation types to shared types
