+++ src/types/recommendations.ts
@@ -59,6 +59,20 @@ export type SyneriseRecommendationsRequest = {
 
 export type SyneriseRecommendationsResult = {
   recommendations: (
-    params: SyneriseRecommendationsRequest
+    params: SyneriseRecommendationsRequest,
   ) => SyneriseRecommendationsResponse;
+};
+
+export type RecommendationViewEvent = {
+  name: "recommendation_view";
+  params: {
+    campaignId: string;
+    correlationId?: string;
+    items: string[];
+  };
+};
+
+export type RecommendationClickEvent = {
+  name: "recommendation_click";
+  params: { campaignId: string; correlationId?: string; item: string };
 };
